+++ src/BlockHorizons/BlockSniper/brush/types/FlattenType.php
@@ -5,8 +5,8 @@ declare(strict_types=1);
 namespace BlockHorizons\BlockSniper\brush\types;
 
 use BlockHorizons\BlockSniper\brush\BaseType;
+use pocketmine\block\Block;
 use pocketmine\block\Flowable;
-use pocketmine\item\Item;
 use pocketmine\level\ChunkManager;
 use pocketmine\math\Vector3;
 use pocketmine\Player;
@@ -29,7 +29,7 @@ class FlattenType extends BaseType{
 	 */
 	public function fill() : \Generator{
 		foreach($this->blocks as $block){
-			if($block->y <= $this->center->y && ($block->getId() === Item::AIR || $block instanceof Flowable)){
+			if($block->y <= $this->center->y && ($block->getId() === Block::AIR || $block instanceof Flowable)){
 				yield $block;
 				$this->putBlock($block, $this->randomBrushBlock());
 			}